refactor: 일부 에러 핸들러 구현
