+++ src/main/java/com/example/satto/global/common/exception/handler/MailHandler.java
@@ -1,0 +1,10 @@
+package com.example.satto.global.common.exception.handler;
+
+import com.example.satto.global.common.code.BaseErrorCode;
+import com.example.satto.global.common.exception.GeneralException;
+
+public class MailHandler extends GeneralException {
+    public MailHandler(BaseErrorCode code) {
+        super(code);
+    }
+}
